joeddhjhiod

+++ Projet/Capteurs en tout genre/Temperature+humidité/src/main.cpp
@@ -6,8 +6,8 @@
 
 Adafruit_AM2320 am2320 = Adafruit_AM2320();
 
-float gettemp();
-float wet();
+int gettemp();
+int wet();
 
 
 void setup() {
@@ -27,17 +27,17 @@ void loop()
   
 }
 
-float gettemp() 
+int gettemp() 
 {
-  float temp = am2320.readTemperature();
+  int temp = am2320.readTemperature();
   Serial.print("Température: "); Serial.println(temp);
   return (temp);
   
 }
 
-float wet() 
+int wet() 
 {
-  float humidity = am2320.readHumidity();
+  int humidity = am2320.readHumidity();
   Serial.print("Humidité: "); Serial.println(humidity);
   return (humidity);
 }
